refactor: Restructure App and HomePage components by moving meeting info and language controls to HomePage. Simplify event listener setup and enhance session management logic. Clean up unused code and improve error handling in language initialization.
